k8s/contratos-inteligentes/requirements/requirements.txt
pandas
google-cloud-bigquery
db_dtypes
psycopg2